+++ contracts/stake/src/contract.rs
@@ -5,6 +5,7 @@ use soroban_sdk::{
 };
 
 use crate::distribution::calc_power;
+use crate::storage::SECONDS_IN_A_DAY;
 use crate::TOKEN_PER_POWER;
 use crate::{
     distribution::{
@@ -168,13 +169,25 @@ impl StakingTrait for Staking {
         lp_token_client.transfer(&sender, &env.current_contract_address(), &tokens);
 
         let mut stakes = get_stakes(&env, &sender);
-        let stake = Stake {
-            stake: tokens,
-            stake_timestamp: ledger.timestamp(),
-        };
+        let mut last_stake = stakes.stakes.last().unwrap_or_default();
+
+        let now = ledger.timestamp();
+
         stakes.total_stake += tokens;
-        // TODO: Discuss: Add implementation to add stake if another is present in +-24h timestamp to avoid
-        // creating multiple stakes the same day
+        // if user bonds again within 12 hours, merge his stakes
+        let stake = if now - (SECONDS_IN_A_DAY / 2) < last_stake.stake_timestamp {
+            last_stake.stake += tokens;
+            last_stake.stake_timestamp = now;
+            // we get rid of the last stake
+            stakes.stakes.pop_back();
+            last_stake
+        } else {
+            Stake {
+                stake: tokens,
+                stake_timestamp: ledger.timestamp(),
+            }
+        };
+        stakes.stakes.push_back(stake);
 
         for distribution_address in get_distributions(&env) {
             let mut distribution = get_distribution(&env, &distribution_address);
@@ -191,7 +204,6 @@ impl StakingTrait for Staking {
             );
         }
 
-        stakes.stakes.push_back(stake);
         save_stakes(&env, &sender, &stakes);
         utils::increase_total_staked(&env, &tokens);