+++ contracts/stake/src/tests/distribution.rs
@@ -6,8 +6,12 @@ use soroban_sdk::{
 use super::setup::{deploy_staking_contract, deploy_token_contract};
 use pretty_assertions::assert_eq;
 
-use crate::msg::{
-    AnnualizedReward, AnnualizedRewardsResponse, WithdrawableReward, WithdrawableRewardsResponse,
+use crate::{
+    msg::{
+        AnnualizedReward, AnnualizedRewardsResponse, WithdrawableReward,
+        WithdrawableRewardsResponse,
+    },
+    tests::setup::{ONE_DAY, ONE_WEEK},
 };
 
 #[test]
@@ -38,16 +42,16 @@ fn add_distribution_and_distribute_reward() {
 
     // bond tokens for user to enable distribution for him
     lp_token.mint(&user, &1000);
-    staking.bond(&user, &1000);
-
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_000;
+        li.timestamp = ONE_DAY;
     });
+
+    staking.bond(&user, &1000);
 
     let reward_duration = 600;
     staking.fund_distribution(
         &admin,
-        &2_000,
+        &ONE_DAY,
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
@@ -60,7 +64,7 @@ fn add_distribution_and_distribute_reward() {
     );
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_600;
+        li.timestamp = ONE_DAY + reward_duration;
     });
     staking.distribute_rewards();
     assert_eq!(
@@ -120,23 +124,20 @@ fn two_distributions() {
 
     // bond tokens for user to enable distribution for him
     lp_token.mint(&user, &1000);
+    env.ledger().with_mut(|li| li.timestamp = ONE_DAY);
     staking.bond(&user, &1000);
-
-    env.ledger().with_mut(|li| {
-        li.timestamp = 2_000;
-    });
 
     let reward_duration = 600;
     staking.fund_distribution(
         &admin,
-        &2_000,
+        &ONE_DAY,
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
     );
     staking.fund_distribution(
         &admin,
-        &2_000,
+        &ONE_DAY,
         &reward_duration,
         &reward_token_2.address,
         &((reward_amount * 2) as i128),
@@ -144,7 +145,7 @@ fn two_distributions() {
 
     // distribute rewards during half time
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_300;
+        li.timestamp += 300;
     });
     staking.distribute_rewards();
     assert_eq!(
@@ -168,7 +169,7 @@ fn two_distributions() {
     assert_eq!(reward_token_2.balance(&user), reward_amount as i128);
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_600;
+        li.timestamp += 600;
     });
     staking.distribute_rewards();
     // first reward token
@@ -245,6 +246,10 @@ fn four_users_with_different_stakes() {
     reward_token.mint(&admin, &(reward_amount as i128));
 
     // bond tokens for users; each user has a different amount staked
+    env.ledger().with_mut(|li| {
+        li.timestamp = ONE_WEEK;
+    });
+
     lp_token.mint(&user, &1000);
     staking.bond(&user, &1000);
     lp_token.mint(&user2, &2000);
@@ -254,21 +259,20 @@ fn four_users_with_different_stakes() {
     lp_token.mint(&user4, &4000);
     staking.bond(&user4, &4000);
 
-    env.ledger().with_mut(|li| {
-        li.timestamp = 2_000;
-    });
+    let eight_days = ONE_WEEK + ONE_DAY;
+    env.ledger().with_mut(|li| li.timestamp = eight_days);
 
     let reward_duration = 600;
     staking.fund_distribution(
         &admin,
-        &2_000,
+        &eight_days,
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
     );
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_600;
+        li.timestamp = eight_days + 600;
     });
     staking.distribute_rewards();
 
@@ -363,23 +367,20 @@ fn two_users_one_starts_after_distribution_begins() {
 
     // first user bonds before distribution started
     lp_token.mint(&user, &1000);
+    env.ledger().with_mut(|li| li.timestamp = ONE_DAY);
     staking.bond(&user, &1000);
-
-    env.ledger().with_mut(|li| {
-        li.timestamp = 2_000;
-    });
 
     let reward_duration = 600;
     staking.fund_distribution(
         &admin,
-        &2_000,
+        &ONE_DAY,
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
     );
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_300;
+        li.timestamp += 300;
     });
     staking.distribute_rewards();
 
@@ -402,7 +403,7 @@ fn two_users_one_starts_after_distribution_begins() {
     staking.bond(&user2, &1000);
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_600;
+        li.timestamp += 300;
     });
     staking.distribute_rewards();
 
@@ -465,21 +466,19 @@ fn two_users_both_bonds_after_distribution_starts() {
     let reward_amount: u128 = 100_000;
     reward_token.mint(&admin, &(reward_amount as i128));
 
-    env.ledger().with_mut(|li| {
-        li.timestamp = 2_000;
-    });
+    env.ledger().with_mut(|li| li.timestamp = ONE_DAY);
 
     let reward_duration = 600;
     staking.fund_distribution(
         &admin,
-        &2_000,
+        &ONE_DAY,
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
     );
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_200;
+        li.timestamp += 200;
     });
     lp_token.mint(&user, &1000);
     staking.bond(&user, &1000);
@@ -502,7 +501,7 @@ fn two_users_both_bonds_after_distribution_starts() {
 
     // user2 starts staking after the distribution has begun
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_400;
+        li.timestamp += 200;
     });
     lp_token.mint(&user2, &1000);
     staking.bond(&user2, &1000);
@@ -534,7 +533,7 @@ fn two_users_both_bonds_after_distribution_starts() {
     );
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_600;
+        li.timestamp += 200;
     });
     staking.distribute_rewards();
 
@@ -765,14 +764,14 @@ fn calculate_apr() {
     reward_token.mint(&admin, &(reward_amount as i128));
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 0;
+        li.timestamp = ONE_DAY;
     });
 
     // whole year of distribution
     let reward_duration = 60 * 60 * 24 * 365;
     staking.fund_distribution(
         &admin,
-        &2_000,
+        &ONE_DAY,
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
@@ -794,6 +793,9 @@ fn calculate_apr() {
 
     // bond tokens for user to enable distribution for him
     lp_token.mint(&user, &1000);
+    env.ledger().with_mut(|li| {
+        li.timestamp += ONE_DAY;
+    });
     staking.bond(&user, &1000);
 
     // 100k rewards distributed for the whole year gives 100% APR
@@ -804,7 +806,7 @@ fn calculate_apr() {
                 &env,
                 AnnualizedReward {
                     asset: reward_token.address.clone(),
-                    amount: String::from_str(&env, "100000")
+                    amount: String::from_str(&env, "100000.975274725274725274")
                 }
             ]
         }
@@ -815,7 +817,7 @@ fn calculate_apr() {
 
     staking.fund_distribution(
         &admin,
-        &2_000,
+        &(2 * &ONE_DAY),
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
@@ -829,7 +831,7 @@ fn calculate_apr() {
                 &env,
                 AnnualizedReward {
                     asset: reward_token.address.clone(),
-                    amount: String::from_str(&env, "150000")
+                    amount: String::from_str(&env, "149727")
                 }
             ]
         }
@@ -891,19 +893,21 @@ fn test_v_phx_vul_010_unbond_breakes_reward_distribution() {
     // bond tokens for user to enable distribution for him
     lp_token.mint(&user_1, &1_000);
     lp_token.mint(&user_2, &1_000);
+
+    env.ledger().with_mut(|li| li.timestamp = ONE_DAY);
     staking.bond(&user_1, &1_000);
     staking.bond(&user_2, &1_000);
     let reward_duration = 10_000;
     staking.fund_distribution(
         &admin,
-        &0,
+        &ONE_DAY,
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
     );
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 2_000;
+        li.timestamp += 2_000;
     });
 
     staking.distribute_rewards();
@@ -914,7 +918,7 @@ fn test_v_phx_vul_010_unbond_breakes_reward_distribution() {
 
     // at the 1/2 of the distribution time, user_1 unbonds
     env.ledger().with_mut(|li| {
-        li.timestamp = 5_000;
+        li.timestamp += 3_000;
     });
     staking.distribute_rewards();
     assert_eq!(
@@ -935,7 +939,7 @@ fn test_v_phx_vul_010_unbond_breakes_reward_distribution() {
             ]
         }
     );
-    staking.unbond(&user_1, &1_000, &0);
+    staking.unbond(&user_1, &1_000, &ONE_DAY);
     assert_eq!(
         staking.query_withdrawable_rewards(&user_1),
         WithdrawableRewardsResponse {
@@ -950,7 +954,7 @@ fn test_v_phx_vul_010_unbond_breakes_reward_distribution() {
     );
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 10_000;
+        li.timestamp += 10_000;
     });
 
     staking.distribute_rewards();
@@ -1007,25 +1011,26 @@ fn test_bond_withdraw_unbond() {
     reward_token.mint(&admin, &(reward_amount as i128));
 
     lp_token.mint(&user, &1_000);
+    env.ledger().with_mut(|li| li.timestamp = ONE_DAY);
     staking.bond(&user, &1_000);
 
     let reward_duration = 10_000;
 
     staking.fund_distribution(
         &admin,
-        &0,
+        &ONE_DAY,
         &reward_duration,
         &reward_token.address,
         &(reward_amount as i128),
     );
 
     env.ledger().with_mut(|li| {
-        li.timestamp = 10_000;
+        li.timestamp = ONE_DAY + reward_duration;
     });
 
     staking.distribute_rewards();
 
-    staking.unbond(&user, &1_000, &0);
+    staking.unbond(&user, &1_000, &ONE_DAY);
 
     assert_eq!(
         staking.query_withdrawable_rewards(&user),
@@ -1133,4 +1138,121 @@ fn panic_when_funding_distribution_with_curve_too_complex() {
         &reward_token.address,
         &1000,
     );
+}
+
+#[test]
+fn one_user_bond_twice_in_a_day_bond_one_more_time_after_a_week_get_rewards() {
+    let env = Env::default();
+    env.mock_all_auths();
+
+    let admin = Address::generate(&env);
+    let user = Address::generate(&env);
+    let manager = Address::generate(&env);
+    let lp_token = deploy_token_contract(&env, &admin);
+    let reward_token = deploy_token_contract(&env, &admin);
+
+    let staking = deploy_staking_contract(
+        &env,
+        admin.clone(),
+        &lp_token.address,
+        &manager,
+        &Address::generate(&env),
+        &50u32,
+    );
+
+    staking.create_distribution_flow(&manager, &reward_token.address);
+
+    let reward_amount: u128 = 100_000;
+    reward_token.mint(&admin, &(reward_amount as i128));
+
+    // bond tokens for user to enable distribution for him
+    lp_token.mint(&user, &3000);
+    env.ledger().with_mut(|li| {
+        li.timestamp = ONE_DAY;
+    });
+
+    staking.fund_distribution(
+        &admin,
+        &ONE_DAY,
+        &(2 * ONE_WEEK),
+        &reward_token.address,
+        &(reward_amount as i128),
+    );
+
+    // first bond for the day
+    staking.bond(&user, &1000);
+
+    staking.distribute_rewards();
+
+    // it's the start of the rewards distribution, so we should have 0 distributed rewards
+    assert_eq!(
+        staking.query_undistributed_rewards(&reward_token.address),
+        reward_amount
+    );
+
+    // user bonds for the second time in a day (12 hours later minus one second)
+    env.ledger()
+        .with_mut(|li| li.timestamp += (ONE_DAY / 2) - 1);
+    staking.bond(&user, &1000);
+
+    assert_eq!(staking.query_staked(&user).stakes.len(), 1);
+
+    env.ledger().with_mut(|li| {
+        li.timestamp += ONE_DAY / 2;
+    });
+
+    staking.distribute_rewards();
+    // distribuion rewards duration is 2 weeks, so after 1 days after starting we should have 1/14 of the rewards
+    // 1/14 out of 100_000 is 7142
+    assert_eq!(
+        staking.query_undistributed_rewards(&reward_token.address),
+        92858
+    );
+    assert_eq!(
+        staking.query_distributed_rewards(&reward_token.address),
+        7142
+    );
+
+    // user bonds for a third time in the middle of the distribution period
+    env.ledger().with_mut(|li| li.timestamp = ONE_WEEK);
+    staking.bond(&user, &1_000);
+
+    env.ledger()
+        .with_mut(|li| li.timestamp = ONE_WEEK + ONE_DAY);
+
+    staking.distribute_rewards();
+
+    // after a week and a day we should have %50 of the rewards
+    assert_eq!(
+        staking.query_undistributed_rewards(&reward_token.address),
+        50_000
+    );
+    assert_eq!(
+        staking.query_distributed_rewards(&reward_token.address),
+        50_000
+    );
+
+    // reward period is over
+    env.ledger()
+        .with_mut(|li| li.timestamp = 2 * ONE_WEEK + ONE_DAY);
+    staking.distribute_rewards();
+
+    assert_eq!(
+        staking.query_withdrawable_rewards(&user),
+        WithdrawableRewardsResponse {
+            rewards: vec![
+                &env,
+                WithdrawableReward {
+                    reward_address: reward_token.address.clone(),
+                    reward_amount: 99999
+                }
+            ]
+        }
+    );
+
+    staking.withdraw_rewards(&user);
+    // it's not actually one whole token that's left, but a fraction of that. I guess we owe this to the
+    // rewards distribution timestamp I'm working with
+    assert_eq!(reward_token.balance(&user), 99999);
+    assert_eq!(reward_token.balance(&staking.address), 1);
 }